Moves flip draw param and fixes screenshot
